prawbot/requirements.txt
python-dotenv
praw
boto3